fix: 클래스 댓글 평점 float 수정

+++ src/apis/class_reviews/dto/create-classReview.input.ts
@@ -1,11 +1,11 @@
-import { Field, InputType, Int } from '@nestjs/graphql';
+import { Field, Float, InputType } from '@nestjs/graphql';
 
 @InputType()
 export class CreateClassReviewInput {
   @Field(() => String)
   content: string;
 
-  @Field(() => Int)
+  @Field(() => Float)
   grade: number;
 
   @Field(() => String)
